+++ wagtaildraftail/client/entities/Link.js
@@ -1,13 +1,6 @@
 import React from 'react';
 import { Entity } from 'draft-js';
-import { ENTITY_TYPE, Icon } from 'draftail';
-
-export function findDocumentEntities(contentBlock, callback) {
-  contentBlock.findEntityRanges((character) => {
-    const entityKey = character.getEntity();
-    return entityKey !== null && Entity.get(entityKey).getType() === ENTITY_TYPE.LINK;
-  }, callback);
-}
+import { Icon } from 'draftail';
 
 const Link = ({ entityKey, children }) => {
   const { url } = Entity.get(entityKey).getData();