+++ CHANGELOG.md
@@ -3,12 +3,16 @@
 > All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/springload/wagtaildraftail/compare/v0.3.0...HEAD)
+## [Unreleased](https://github.com/springload/wagtaildraftail/compare/v0.4.0...HEAD)
+
+...
+
+## [[v0.4.0]](https://github.com/springload/draftjs_exporter/releases/tag/v0.4.0)
 
 ### Add
 
 - Equality check for DraftText nodes.
-- Default entity strategy.
+- Add default entity strategy.
 
 ### Changed